[WIP] Move proxy homepage to a generic function and load funding finder with it
