+++ controllers/toplevel/index.js
@@ -14,6 +14,7 @@ const app = require('../../server');
 const routeStatic = require('../utils/routeStatic');
 const regions = require('../../config/content/regions.json');
 const models = require('../../models/index');
+const proxyLegacy = require('../../modules/proxy');
 
 const robots = require('../../config/app/robots.json');
 // block everything on non-prod envs
@@ -99,81 +100,11 @@ const newHomepage = (req, res) => {
     }
 };
 
-// @TODO cache this page as it's very slow to return
-// main issue is the static assets (which cloudfront doesn't cache)
 const oldHomepage = (req, res) => {
-    // don't cache this page!
-    res.cacheControl = { maxAge: 0 };
-
-    // work out if we need to serve english/welsh page
-    let localePath = req.i18n.getLocale() === 'cy' ? config.get('i18n.urlPrefix.cy') : '';
-
-    return rp({
-        url: legacyUrl + localePath,
-        strictSSL: false,
-        jar: true,
-        resolveWithFullResponse: true
-    })
-        .then(response => {
-            let body = response.body;
-            // convert all links in the document to be absolute
-            // (only really useful on non-prod envs)
-            body = absolution(body, 'https://www.biglotteryfund.org.uk');
-
-            // fix meta tags in HTML which use the wrong CNAME
-            body = body.replace(/wwwlegacy/g, 'www');
-
-            // parse the DOM
-            const dom = new JSDOM(body);
-
-            const form = dom.window.document.getElementById('form1');
-            if (form) {
-                const newAction = form.getAttribute('action').replace('https://www.biglotteryfund.org.uk/', '/');
-                form.setAttribute('action', newAction);
-            }
-
-            // are we in an A/B test?
-            if (res.locals.ab) {
-                // create GA snippet for tracking experiment
-                const gaCode = `
-                <script src="//www.google-analytics.com/cx/api.js"></script>
-                <script>
-                    (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-                            (i[r].q=i[r].q||[]).push(arguments);},i[r].l=1*new Date();a=s.createElement(o),
-                        m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m);
-                    })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-                    ga('create', '${config.get('googleAnalyticsCode')}', {
-                        'cookieDomain': 'none'
-                    });
-                    ga('set', 'expId', '${res.locals.ab.id}');
-                    ga('set', 'expVar', ${res.locals.ab.variantId});
-                    cxApi.setChosenVariation(${res.locals.ab.variantId}, '${res.locals.ab.id}');
-                    ga('send', 'pageview');
-                </script>`;
-
-                // insert GA experiment code into the page
-                const script = dom.window.document.createElement('div');
-                script.innerHTML = gaCode;
-                dom.window.document.body.appendChild(script);
-
-                // try to kill the google tag manager (useful for non-prod envs)
-                // @TODO kill the noscript too?
-                // @TODO don't do this on prod?
-                const scripts = dom.window.document.scripts;
-                let gtm = [].find.call(scripts, s => s.innerHTML.indexOf('www.googletagmanager.com/gtm.js') !== -1);
-                if (gtm) {
-                    gtm.innerHTML = '';
-                }
-            }
-            res.set('X-BLF-Legacy', true);
-            res.send(dom.serialize());
-        })
-        .catch(error => {
-            // we failed to fetch from the proxy, redirect to new
-            console.log('Error fetching legacy site', error);
-            res.redirect('/home');
-        });
+    return proxyLegacy.proxyLegacyPage(req, res);
 };
+
+router.get('/funding/funding-finder', proxyLegacy.proxyLegacyPage);
 
 const oldHomepagePost = (req, res) => {
     res.cacheControl = { maxAge: 0 };
